#235: Attempt at Gutenberg sidebar...
First impressions: I don't like Gutenberg: It's not ready, it's not easy, it desolates and neglects current developers, it'll probably be the end of WordPress' FOSS developer stream (thus permeating to its own end), it absolutely destroys your dashboard performance (thus flow of operation), it's prone to unrecoverable and completely annihilating crashes, and the documentation leaves a lot to be desired. It's no wonder that even the big boys don't integrate their settings correctly whilst being full-on into the project.

That said, I was able to find a few bugs implementing this.